OMF module reorganization
