+++ src/read/mod.rs
@@ -355,6 +355,8 @@ impl FileKind {
             #[cfg(feature = "omf")]
             [0x80, ..] | [0x82, ..] => {
                 // Check if it's a valid OMF record type
+                // TODO this is tautological, 0x80 and 0x82 are valid OMF record types
+                // how can we check better?
                 if crate::omf::is_omf_record_type(magic[0]) {
                     FileKind::Omf
                 } else {
