+++ src/read/omf/section.rs
@@ -1,24 +1,29 @@
-//! OMF section implementation.
-
 use alloc::borrow::Cow;
-use alloc::vec;
+use alloc::{vec, vec::Vec};
 use core::str;
 
-use crate::{
-    read::{
-        self, CompressedData, CompressedFileRange, Error, ObjectSection, ReadRef, RelocationMap,
-        Result, SectionFlags, SectionIndex, SectionKind,
-    },
-    ComdatKind, ObjectComdat, SymbolIndex,
+use crate::read::{
+    self, CompressedData, CompressedFileRange, Error, ObjectSection, ReadRef, RelocationMap,
+    Result, SectionFlags, SectionIndex, SectionKind,
 };
 
-use super::{relocation::OmfRelocationIterator, OmfDataChunk, OmfFile, OmfSegment};
+use super::{expand_lidata_block, OmfDataChunk, OmfFile, OmfRelocationIterator, OmfSegment};
 
 /// A section in an OMF file.
 #[derive(Debug)]
 pub struct OmfSection<'data, 'file, R: ReadRef<'data>> {
     pub(super) file: &'file OmfFile<'data, R>,
     pub(super) index: usize,
+}
+
+/// An OMF group definition
+#[derive(Debug, Clone)]
+#[allow(unused)]
+pub(super) struct OmfGroup {
+    /// Group name index (into names table)
+    pub(super) name_index: u16,
+    /// Segment indices in this group
+    pub(super) segments: Vec<u16>,
 }
 
 impl<'data, 'file, R: ReadRef<'data>> OmfSection<'data, 'file, R> {
@@ -153,7 +158,7 @@ impl<'data, 'file, R: ReadRef<'data>> ObjectSection<'data> for OmfSection<'data,
                 }
                 OmfDataChunk::Iterated(lidata) => {
                     // LIDATA needs expansion
-                    if let Ok(expanded) = self.file.expand_lidata_block(lidata) {
+                    if let Ok(expanded) = expand_lidata_block(lidata) {
                         let end = start + expanded.len();
                         if end <= result.len() {
                             result[start..end].copy_from_slice(&expanded);
@@ -240,100 +245,6 @@ impl<'data, 'file, R: ReadRef<'data>> Iterator for OmfSectionIterator<'data, 'fi
             };
             self.index += 1;
             Some(section)
-        } else {
-            None
-        }
-    }
-}
-
-/// A COMDAT section in an OMF file.
-#[derive(Debug)]
-pub struct OmfComdat<'data, 'file, R: ReadRef<'data>> {
-    file: &'file OmfFile<'data, R>,
-    index: usize,
-    _phantom: core::marker::PhantomData<&'data ()>,
-}
-
-impl<'data, 'file, R: ReadRef<'data>> read::private::Sealed for OmfComdat<'data, 'file, R> {}
-
-impl<'data, 'file, R: ReadRef<'data>> ObjectComdat<'data> for OmfComdat<'data, 'file, R> {
-    type SectionIterator = OmfComdatSectionIterator<'data, 'file, R>;
-
-    fn kind(&self) -> ComdatKind {
-        let comdat = &self.file.comdats[self.index];
-        match comdat.selection {
-            super::OmfComdatSelection::Explicit => ComdatKind::NoDuplicates,
-            super::OmfComdatSelection::UseAny => ComdatKind::Any,
-            super::OmfComdatSelection::SameSize => ComdatKind::SameSize,
-            super::OmfComdatSelection::ExactMatch => ComdatKind::ExactMatch,
-        }
-    }
-
-    fn symbol(&self) -> SymbolIndex {
-        // COMDAT symbols don't have a direct symbol index in OMF
-        SymbolIndex(usize::MAX)
-    }
-
-    fn name_bytes(&self) -> Result<&'data [u8]> {
-        let comdat = &self.file.comdats[self.index];
-        Ok(comdat.name)
-    }
-
-    fn name(&self) -> Result<&'data str> {
-        let comdat = &self.file.comdats[self.index];
-        core::str::from_utf8(comdat.name).map_err(|_| Error("Invalid UTF-8 in COMDAT name"))
-    }
-
-    fn sections(&self) -> Self::SectionIterator {
-        let comdat = &self.file.comdats[self.index];
-        OmfComdatSectionIterator {
-            segment_index: (comdat.segment_index as usize).checked_sub(1),
-            returned: false,
-            _phantom: core::marker::PhantomData,
-        }
-    }
-}
-
-/// An iterator over COMDAT sections.
-#[derive(Debug)]
-pub struct OmfComdatIterator<'data, 'file, R: ReadRef<'data>> {
-    pub(super) file: &'file OmfFile<'data, R>,
-    pub(super) index: usize,
-}
-
-impl<'data, 'file, R: ReadRef<'data>> Iterator for OmfComdatIterator<'data, 'file, R> {
-    type Item = OmfComdat<'data, 'file, R>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.index < self.file.comdats.len() {
-            let comdat = OmfComdat {
-                file: self.file,
-                index: self.index,
-                _phantom: core::marker::PhantomData,
-            };
-            self.index += 1;
-            Some(comdat)
-        } else {
-            None
-        }
-    }
-}
-
-/// An iterator over sections in a COMDAT.
-#[derive(Debug)]
-pub struct OmfComdatSectionIterator<'data, 'file, R: ReadRef<'data>> {
-    segment_index: Option<usize>,
-    returned: bool,
-    _phantom: core::marker::PhantomData<(&'data (), &'file (), R)>,
-}
-
-impl<'data, 'file, R: ReadRef<'data>> Iterator for OmfComdatSectionIterator<'data, 'file, R> {
-    type Item = SectionIndex;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if !self.returned {
-            self.returned = true;
-            self.segment_index.map(|idx| SectionIndex(idx + 1))
         } else {
             None
         }
